Reorganize mock_test sai_wraps
